Rename status_verbosity.hpp to status.hpp
This file will receive additional functionality in the next commit.
This commit contains the mechanical preparation.
