+++ libvast/src/system/accountant.cpp
@@ -20,7 +20,7 @@
 #include "vast/error.hpp"
 #include "vast/logger.hpp"
 #include "vast/system/report.hpp"
-#include "vast/system/status_verbosity.hpp"
+#include "vast/system/status.hpp"
 #include "vast/table_slice.hpp"
 #include "vast/table_slice_builder_factory.hpp"
 #include "vast/time.hpp"
