+++ libvast/src/system/datagram_source.cpp
@@ -21,7 +21,7 @@
 #include "vast/expression_visitors.hpp"
 #include "vast/logger.hpp"
 #include "vast/schema.hpp"
-#include "vast/system/status_verbosity.hpp"
+#include "vast/system/status.hpp"
 #include "vast/system/transformer.hpp"
 #include "vast/table_slice.hpp"
