+++ libvast/src/system/posix_filesystem.cpp
@@ -12,7 +12,7 @@
 #include "vast/detail/assert.hpp"
 #include "vast/io/read.hpp"
 #include "vast/io/save.hpp"
-#include "vast/system/status_verbosity.hpp"
+#include "vast/system/status.hpp"
 
 #include <caf/config_value.hpp>
 #include <caf/dictionary.hpp>
