+++ libvast/src/system/transformer.cpp
