+++ libvast/src/system/type_registry.cpp
@@ -17,7 +17,7 @@
 #include "vast/io/save.hpp"
 #include "vast/logger.hpp"
 #include "vast/system/report.hpp"
-#include "vast/system/status_verbosity.hpp"
+#include "vast/system/status.hpp"
 #include "vast/table_slice.hpp"
 
 #include <caf/attach_stream_sink.hpp>
