+++ libvast/vast/system/status.hpp
